Cleans up files adds temporary providers goog - fb - twitter
